refactor: copy-build-assets をTypeScript化し tsx で実行するよう変更
